+++ examples/web/src/__tests__/authentication-log-in-and-out.test.js
@@ -1,0 +1,229 @@
+import * as Realm from "realm-web";
+import { APP_ID } from "../realm.config.json";
+
+const app = new Realm.App({ id: APP_ID });
+
+describe("Log in user", () => {
+  afterEach(async () => {
+    await app?.currentUser?.logOut();
+  });
+  test("Anonymous log in", async () => {
+    // :snippet-start: anon-auth
+    async function loginAnonymous() {
+      // Create an anonymous credential
+      const credentials = Realm.Credentials.anonymous();
+      try {
+        // Authenticate the user
+        const user = await app.logIn(credentials);
+        // `App.currentUser` updates to match the logged in user
+        console.assert(user.id === app.currentUser.id);
+        return user;
+      } catch (err) {
+        console.error("Failed to log in", err);
+      }
+    }
+    const user = await loginAnonymous();
+    console.log("Successfully logged in!", user);
+
+    // :snippet-end:
+    expect(app.currentUser.isLoggedIn).toBe(true);
+    expect(app.currentUser?.id).toBe(user.id);
+  });
+  test("Email/password user", async () => {
+    // :snippet-start: email-password-auth
+    async function loginEmailPassword(email, password) {
+      // Create an anonymous credential
+      const credentials = Realm.Credentials.emailPassword(email, password);
+      try {
+        // Authenticate the user
+        const user = await app.logIn(credentials);
+        // `App.currentUser` updates to match the logged in user
+        console.assert(user.id === app.currentUser.id);
+        return user;
+      } catch (err) {
+        console.error("Failed to log in", err);
+      }
+    }
+
+    const user = await loginEmailPassword("joe.jasper@example.com", "passw0rd");
+    console.log("Successfully logged in!", user);
+    // :snippet-end:
+    expect(app.currentUser.isLoggedIn).toBe(true);
+    expect(app.currentUser?.id).toBe(user.id);
+  });
+  test("API key", async () => {
+    const baseUser = await app.logIn(Realm.Credentials.anonymous());
+    const REALM_API_KEY = baseUser.apiKeys.create("myKey");
+    await baseUser.logOut();
+    // :snippet-start: api-key-auth
+    async function loginApiKey(apiKey) {
+      // Create an API Key credential
+      const credentials = Realm.Credentials.apiKey(apiKey);
+      try {
+        // Authenticate the user
+        const user = await app.logIn(credentials);
+        // `App.currentUser` updates to match the logged in user
+        console.assert(user.id === app.currentUser.id);
+        return user;
+      } catch (err) {
+        console.error("Failed to log in", err);
+      }
+    }
+    const user = await loginApiKey(REALM_API_KEY); // add previously generated API key
+    console.log("Successfully logged in!", user);
+    // :snippet-end:
+    expect(app.currentUser.isLoggedIn).toBe(true);
+    expect(app.currentUser?.id).toBe(user.id);
+  });
+  // TODO
+  test.skip("Custom function", () => {
+    // :snippet-start: custom-function-auth
+    async function loginCustomFunction(payload) {
+      // Create a Custom Function credential
+      const credentials = Realm.Credentials.function(payload);
+      try {
+        // Authenticate the user
+        const user = await app.logIn(credentials);
+        // `App.currentUser` updates to match the logged in user
+        console.assert(user.id === app.currentUser.id);
+        return user;
+      } catch (err) {
+        console.error("Failed to log in", err);
+      }
+    }
+    loginCustomFunction({ username: "mongolover" }).then((user) => {
+      console.log("Successfully logged in!", user);
+    });
+    // :snippet-end:
+  });
+  // TODO
+  test.skip("Custom JWT", () => {
+    // :snippet-start: custom-jwt
+    async function loginCustomJwt(jwt) {
+      // Create a Custom JWT credential
+      const credentials = Realm.Credentials.jwt(jwt);
+      try {
+        // Authenticate the user
+        const user = await app.logIn(credentials);
+        // `App.currentUser` updates to match the logged in user
+        console.assert(user.id === app.currentUser.id);
+        return user;
+      } catch (err) {
+        console.error("Failed to log in", err);
+      }
+    }
+    loginCustomJwt("eyJ0eXAi...Q3NJmnU8oP3YkZ8").then((user) => {
+      console.log("Successfully logged in!", user);
+    });
+    // :snippet-end:
+  });
+  // TODO
+  describe.skip("Facebook OAuth", () => {
+    test("Built-in Facebook OAuth", () => {
+      // :snippet-start: builtin-facebook-oauth
+      // The redirect URI should be on the same domain as this app and
+      // specified in the auth provider configuration.
+      const redirectUri = "https://app.example.com/handleOAuthLogin";
+      const credentials = Realm.Credentials.facebook(redirectUri);
+
+      // Calling logIn() opens a Facebook authentication screen in a new window.
+      app.logIn(credentials).then((user) => {
+        // The logIn() promise will not resolve until you call `handleAuthRedirect()`
+        // from the new window after the user has successfully authenticated.
+        console.log(`Logged in with id: ${user.id}`);
+      });
+
+      // When the user is redirected back to your app, handle the redirect to
+      // save the user's access token and close the redirect window. This
+      // returns focus to the original application window and automatically
+      // logs the user in.
+      Realm.handleAuthRedirect();
+      // :snippet-end:
+    });
+    test("Facebook SDK OAuth", () => {
+      // :snippet-start: facebook-sdk-oauth
+      // Get the access token from the Facebook SDK
+      const { accessToken } = FB.getAuthResponse();
+      // Define credentials with the access token from the Facebook SDK
+      const credentials = Realm.Credentials.facebook(accessToken);
+      // Log the user in to your app
+      app.logIn(credentials).then((user) => {
+        console.log(`Logged in with id: ${user.id}`);
+      });
+      // :snippet-end:
+    });
+  });
+  // TODO
+  describe.skip("Google OAuth", () => {
+    test("Built-in Google OAuth", () => {
+      // :snippet-start: builtin-google-oauth
+      // TODO: add abstracted example
+      // :snippet-end:
+    });
+    test("Google Onetap OAuth", () => {
+      // :snippet-start: google-onetap-oauth
+      // TODO: Add abstracted example
+      // :snippet-end:
+    });
+  });
+  // TODO
+  describe.skip("Apple OAuth", () => {
+    test("Built-in Apple OAuth", () => {
+      // :snippet-start: builtin-apple-oauth
+      // The redirect URI should be on the same domain as this app and
+      // specified in the auth provider configuration.
+      const redirectUri = "https://app.example.com/handleOAuthLogin";
+      const credentials = Realm.Credentials.apple(redirectUri);
+
+      // Calling logIn() opens an Apple authentication screen in a new window.
+      app.logIn(credentials).then((user) => {
+        // The logIn() promise will not resolve until you call `handleAuthRedirect()`
+        // from the new window after the user has successfully authenticated.
+        console.log(`Logged in with id: ${user.id}`);
+      });
+
+      // When the user is redirected back to your app, handle the redirect to
+      // save the user's access token and close the redirect window. This
+      // returns focus to the original application window and automatically
+      // logs the user in.
+      Realm.handleAuthRedirect();
+      // :snippet-end:
+    });
+    test("Apple SDK OAuth", () => {
+      // :snippet-start: apple-sdk-oauth
+      // Get the ID token from the Apple SDK
+      AppleID.auth
+        .signIn()
+        .then(({ id_token }) => {
+          // Define credentials with the ID token from the Apple SDK
+          const credentials = Realm.Credentials.apple(id_token);
+          // Log the user in to your app
+          return app.logIn(credentials);
+        })
+        .then((user) => {
+          console.log(`Logged in with id: ${user.id}`);
+        });
+      // :snippet-end:
+    });
+  });
+});
+describe("Log out user", () => {
+  test("Log out current user", async () => {
+    const user = await app.logIn(Realm.Credentials.anonymous());
+    // :snippet-start: log-out-current-user
+    await app.currentUser.logOut();
+    // :snippet-end:
+    expect(user.isLoggedIn).toBe(false);
+    expect(app.currentUser.isLoggedIn).toBe(false);
+  });
+  test("Log out specific user", async () => {
+    const user1 = await app.logIn(Realm.Credentials.anonymous());
+    const user2 = await app.logIn(Realm.Credentials.anonymous());
+    const user3 = await app.logIn(Realm.Credentials.anonymous());
+    // :snippet-start: log-out-specific-user
+    await app.allUsers[2].logOut();
+    // :snippet-end
+    await user1.logOut();
+    await user2.logOut();
+  });
+});
